+++ CHANGELOG.md
@@ -10,8 +10,12 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 ### Added
 - Multi-cursor support
-- New command `extension.deleteWordRight` with keyboard shortcut <kbd>Ctrl+Delete</kbd>
-- New command `extension.deleteWordLeft` with keyboard shortcut <kbd>Ctrl+Backspace</kbd>
+- New command `japaneseWordHandler.deleteWordRight` with keyboard shortcut <kbd>Ctrl+Delete</kbd>
+- New command `japaneseWordHandler.deleteWordLeft` with keyboard shortcut <kbd>Ctrl+Backspace</kbd>
+
+### Changed
+- Rename command naming convension such as `extension.cursorNextWordEndJa` to
+  `japaneseWordHandler.cursorWordEndRight`
 
 ## [1.0.0] - 2018-05-21
 ### Added